fix: 카테고리 최종본 기준으로 변경
