--- features/post/basicTopics.ts
@@ -1,90 +1,0 @@
-export const basicTopics = {
-  공부: [
-    "오늘 공부한 내용 정리",
-    "이해가 어려웠던 부분",
-    "외워야 할 핵심 요약",
-    "공부 진행 상황",
-    "내일 공부 계획",
-    "오늘 공부하며 든 생각",
-  ],
-  운동: [
-    "오늘 운동한 내용 정리",
-    "운동 중 느낀 난이도",
-    "몸 상태와 컨디션",
-    "변화 기록",
-    "다음 운동 계획",
-    "운동하며 떠오른 생각",
-  ],
-  커리어: [
-    "오늘 진행한 주요 업무",
-    "현재 커리어에 대한 생각",
-    "업무 중 새로 배운 점",
-    "어려웠던 문제와 원인",
-    "해결 방법",
-    "향후 커리어 계획",
-  ],
-  독서: [
-    "오늘 읽은 책과 저자",
-    "기억에 남는 문장",
-    "오늘 읽은 분량",
-    "책에서 얻은 인사이트",
-    "읽으며 든 생각",
-    "다음 독서 계획",
-  ],
-  "문화/취미": [
-    "오늘 한 활동",
-    "오늘의 인사이트 기록",
-    "만족스러웠던 포인트",
-    "아쉬웠던 부분",
-    "개선하고 싶은 점",
-    "다음 계획",
-  ],
-  "투자/머니로그": [
-    "오늘 투자한 내용",
-    "오늘 소비한 내용",
-    "수익·손실 요약",
-    "새로 얻은 인사이트",
-    "다음 투자 계획",
-    "마인드셋 기록",
-  ],
-  "챌린지/다짐": [
-    "오늘 지킨 다짐",
-    "성공·실패 요인",
-    "발전하고 있는 점",
-    "개선할 부분",
-    "내일 루틴 계획",
-    "의지를 다지는 한 마디",
-  ],
-  패션: [
-    "오늘 착장 요약",
-    "오늘의 포인트 요소",
-    "마음에 들었던 점",
-    "개선하고 싶은 점",
-    "다음 착장 계획",
-    "새로운 패션 인사이트",
-  ],
-  식단: [
-    "오늘의 식단 요약",
-    "칼로리 / 영양 밸런스",
-    "컨디션 변화",
-    "만족스러운 점",
-    "아쉬웠던 점",
-    "내일 식단 계획",
-  ],
-  "가족/친구": [
-    "오늘 함께한 시간",
-    "즐거웠던 점",
-    "아쉬웠던 점",
-    "앞으로 노력해야할 것",
-    "함께 하며 떠오른 생각",
-    "다음 모임 계획",
-  ],
-  "생각/일기": [
-    "오늘 정리한 생각 하나",
-    "감사했던 일",
-    "놓치고 싶지 않은 관점",
-    "힘들었던 순간",
-    "극복 방법",
-    "내일의 마음가짐",
-  ],
-};
