+++ types/Categories.ts
@@ -1,26 +1,126 @@
-export const categories: Category[] = [
-  "공부",
-  "운동",
-  "커리어",
-  "독서",
-  "문화/취미",
-  "투자/머니로그",
-  "챌린지/다짐",
-  "패션",
-  "식단",
-  "가족/친구",
-  "생각/일기",
-];
+export const basicTopics = {
+  공부: [
+    "공부한 내용 정리",
+    "이해가 어려웠던 부분",
+    "외워야 할 핵심 요약",
+    "공부 진행 상황",
+    "다음 공부 계획",
+    "공부하며 든 생각",
+  ],
+  운동: [
+    "운동한 내용 정리",
+    "운동 중 느낀 난이도",
+    "몸 상태와 컨디션",
+    "변화 기록",
+    "다음 운동 계획",
+    "운동하며 떠오른 생각",
+  ],
+  커리어: [
+    "진행한 주요 업무",
+    "현재 커리어에 대한 생각",
+    "업무 중 새로 배운 점",
+    "어려웠던 문제와 원인",
+    "해결 방법",
+    "향후 커리어 계획",
+  ],
+  "문화/예술": [
+    "오늘의 내용 정리",
+    "인사이트 기록",
+    "만족스러웠던 포인트",
+    "아쉬웠던 부분",
+    "가장 기억에 남는 순간",
+    "다음 계획",
+  ],
+  "루틴/다짐": [
+    "루틴/다짐 기록",
+    "성공·실패 요인",
+    "발전하고 있는 점",
+    "개선할 부분",
+    "의지를 다지는 한 마디",
+    "앞으로의 루틴 계획",
+  ],
+  독서: [
+    "읽은 책과 저자",
+    "읽은 분량",
+    "기억에 남는 문장",
+    "책에서 얻은 인사이트",
+    "읽으며 든 생각",
+    "다음 독서 계획",
+  ],
+  "투자/머니로그": [
+    "투자 기록 정리",
+    "소비 내역 정리",
+    "수익·손실 요약",
+    "새로 얻은 인사이트",
+    "의지를 다지는 한 마디",
+    "앞으로의 계획",
+  ],
+  취미: [
+    "오늘 한 활동",
+    "만족스러웠던 포인트",
+    "발전하고 있는 점",
+    "개선할 부분",
+    "의지를 다지는 한 마디",
+    "앞으로의 계획",
+  ],
+  여행: [
+    "여행 기록 정리",
+    "즐거웠던 점",
+    "아쉬웠던 점",
+    "가장 기억에 남는 순간",
+    "여행하며 든 생각",
+    "다음 여행 계획",
+  ],
+  식단: [
+    "식단 요약",
+    "칼로리 / 영양 밸런스",
+    "컨디션 변화",
+    "만족스러운 점",
+    "아쉬웠던 점",
+    "내일 식단 계획",
+  ],
+  패션: [
+    "착장 요약",
+    "메인 포인트 요소",
+    "마음에 들었던 점",
+    "개선하고 싶은 점",
+    "새로운 패션 인사이트",
+    "다음 착장 계획",
+  ],
+  종교: [
+    "오늘의 하루",
+    "말씀 기록",
+    "감사한 점",
+    "삶의 변화 기록",
+    "기도 제목",
+    "다음 계획",
+  ],
+  반려동물: [
+    "함께한 하루 기록",
+    "오늘의 식사",
+    "건강 상태와 컨디션",
+    "함께 하며 든 생각",
+    "변화 기록",
+    "내일의 계획",
+  ],
+  "가족/모임": [
+    "함께한 시간 요약",
+    "즐거웠던 점",
+    "아쉬웠던 점",
+    "앞으로 함께하고 싶은 것",
+    "함께 하며 떠오른 생각",
+    "다음 계획",
+  ],
+  "생각/일기": [
+    "오늘 정리한 생각 하나",
+    "감사했던 일",
+    "놓치고 싶지 않은 관점",
+    "힘들었던 순간",
+    "극복 방법",
+    "내일의 마음가짐",
+  ],
+} as const;
 
-export type Category =
-  | "공부"
-  | "운동"
-  | "커리어"
-  | "독서"
-  | "문화/취미"
-  | "투자/머니로그"
-  | "챌린지/다짐"
-  | "패션"
-  | "식단"
-  | "가족/친구"
-  | "생각/일기";
+export type Category = keyof typeof basicTopics;
+
+export const categories: Category[] = Object.keys(basicTopics) as Category[];
